refactor(speculative): share grad-accum + LR utils across EAGLE/DFlash recipes
EAGLE-1/2, EAGLE-3, and DFlash each hand-rolled byte-identical copies of the
optimizer-steps-per-epoch ceil, the DDP no_sync gradient-sync decision, and the
warmup + cosine LR lambda. A fix to one was easy to miss in the others (DFlash
had already drifted to an inline copy of the sync logic).

Extract the three into nemo_automodel/recipes/llm/_spec_train_utils.py and have
all three recipes import them. Behavior is unchanged: DFlash's inline sync was
equivalent (its num_batches is always finite), so it now calls the shared
should_sync_grads. The shared functions are unit-tested in one place; the
duplicate per-recipe helper tests are consolidated there (the EAGLE-3 grad-accum
file was purely those and is removed; EAGLE-1 keeps its DDP integration tests).
