+++ tests/unit_tests/recipes/llm/test_train_eagle1_grad_accum.py
@@ -37,86 +37,7 @@ import torch.distributed as dist
 import torch.nn as nn
 from torch.nn.parallel import DistributedDataParallel
 
-from nemo_automodel.recipes.llm.train_eagle1 import (
-    TrainEagle1Recipe,
-    _optim_steps_per_epoch,
-    _should_sync_grads,
-)
-
-
-@pytest.mark.parametrize(
-    "num_batches,accum,expected",
-    [
-        (10, 1, 10),
-        (10, 2, 5),
-        (10, 3, 4),  # 3 full windows + 1 trailing micro-batch -> 4 steps
-        (10, 4, 3),  # 2 full windows + 2 trailing -> 3 steps
-        (1, 4, 1),  # entire epoch is one trailing flush
-        (4, 4, 1),
-        (5, 4, 2),
-        (0, 4, 0),  # iterable dataloader / no length
-    ],
-)
-def test_optim_steps_per_epoch_uses_ceil_division(num_batches, accum, expected):
-    assert _optim_steps_per_epoch(num_batches, accum) == expected
-
-
-def test_optim_steps_per_epoch_handles_invalid_inputs():
-    assert _optim_steps_per_epoch(0, 1) == 0
-    assert _optim_steps_per_epoch(-1, 4) == 0
-    assert _optim_steps_per_epoch(10, 0) == 0
-    assert _optim_steps_per_epoch(10, -1) == 0
-
-
-# ---------------------------------------------------------------------------
-# DDP gradient-sync decision (no_sync)
-# ---------------------------------------------------------------------------
-
-
-def _sync(pending, batch_idx, *, accum=4, batches_per_epoch=10, is_ddp=True):
-    return _should_sync_grads(
-        pending_micro_batches=pending,
-        grad_accumulation_steps=accum,
-        batch_idx=batch_idx,
-        batches_per_epoch=batches_per_epoch,
-        is_ddp=is_ddp,
-    )
-
-
-def test_should_sync_always_true_without_ddp():
-    # Single process: nothing to all-reduce, so every step "syncs" (no_sync is
-    # never entered) regardless of window position or batch index.
-    for pending in range(4):
-        assert _sync(pending, batch_idx=0, is_ddp=False) is True
-
-
-def test_should_sync_only_on_window_close_under_ddp():
-    # accum=4: interior micro-batches defer the all-reduce, the 4th closes it.
-    assert _sync(0, batch_idx=0) is False
-    assert _sync(1, batch_idx=1) is False
-    assert _sync(2, batch_idx=2) is False
-    assert _sync(3, batch_idx=3) is True  # pending+1 == accum -> window closer
-
-
-def test_should_sync_on_epoch_final_batch_even_mid_window():
-    # The trailing-flush step consumes the last batch's grads, so it must sync
-    # even though the window is not full (batch_idx == batches_per_epoch - 1).
-    assert _sync(0, batch_idx=9, batches_per_epoch=10) is True
-    assert _sync(1, batch_idx=9, batches_per_epoch=10) is True
-
-
-def test_should_sync_every_step_when_length_unknown():
-    # IterableDataset (len unknown): we cannot identify the final batch, so a
-    # trailing window could step on un-synced grads -- sync every step instead.
-    for pending in range(4):
-        assert _sync(pending, batch_idx=pending, batches_per_epoch=None) is True
-
-
-def test_should_sync_every_step_when_accum_is_one():
-    # No accumulation: each batch closes its own window -> always sync.
-    for batch_idx in range(5):
-        assert _sync(0, batch_idx=batch_idx, accum=1) is True
-
+from nemo_automodel.recipes.llm.train_eagle1 import TrainEagle1Recipe
 
 # ---------------------------------------------------------------------------
 # Trailing-flush gradient rescale
